Extract methods from NodeCreateHandler to AddressBookValidator
