+++ hedera-node/hedera-addressbook-service-impl/src/main/java/com/hedera/node/app/service/addressbook/impl/validators/AddressBookValidator.java
@@ -1,0 +1,121 @@
+/*
+ * Copyright (C) 2024 Hedera Hashgraph, LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.hedera.node.app.service.addressbook.impl.validators;
+
+import static com.hedera.hapi.node.base.ResponseCodeEnum.FQDN_SIZE_TOO_LARGE;
+import static com.hedera.hapi.node.base.ResponseCodeEnum.GOSSIP_ENDPOINT_CANNOT_HAVE_FQDN;
+import static com.hedera.hapi.node.base.ResponseCodeEnum.INVALID_ENDPOINT;
+import static com.hedera.hapi.node.base.ResponseCodeEnum.INVALID_GOSSIP_ENDPOINT;
+import static com.hedera.hapi.node.base.ResponseCodeEnum.INVALID_NODE_DESCRIPTION;
+import static com.hedera.hapi.node.base.ResponseCodeEnum.INVALID_SERVICE_ENDPOINT;
+import static com.hedera.hapi.node.base.ResponseCodeEnum.IP_FQDN_CANNOT_BE_SET_FOR_SAME_ENDPOINT;
+import static com.hedera.node.app.spi.workflows.HandleException.validateFalse;
+
+import com.hedera.hapi.node.base.ServiceEndpoint;
+import com.hedera.node.config.data.NodesConfig;
+import edu.umd.cs.findbugs.annotations.NonNull;
+import edu.umd.cs.findbugs.annotations.Nullable;
+import java.nio.charset.StandardCharsets;
+import java.util.List;
+import javax.inject.Inject;
+
+public class AddressBookValidator {
+    /**
+     * Default constructor for injection.
+     */
+    @Inject
+    public AddressBookValidator() {
+        // Dagger2
+    }
+
+    /**
+     * Validates the node description.
+     *
+     * @param description The description to validate
+     * @param nodesConfig The nodes configuration
+     */
+    public void validateDescription(@Nullable final String description, @NonNull final NodesConfig nodesConfig) {
+        if (description == null || description.isEmpty()) {
+            return;
+        }
+        final var raw = description.getBytes(StandardCharsets.UTF_8);
+        final var maxUtf8Bytes = nodesConfig.nodeMaxDescriptionUtf8Bytes();
+        validateFalse(raw.length > maxUtf8Bytes, INVALID_NODE_DESCRIPTION);
+        validateFalse(containsZeroByte(raw), INVALID_NODE_DESCRIPTION);
+    }
+
+    private boolean containsZeroByte(@NonNull final byte[] bytes) {
+        boolean ret = false;
+        for (final byte b : bytes) {
+            if (b == 0) {
+                ret = true;
+                break;
+            }
+        }
+        return ret;
+    }
+
+    /**
+     * Validates the gossip endpoint.
+     *
+     * @param endpointLst The list of GossipEndpoint to validate
+     * @param nodesConfig The nodes configuration
+     */
+    public void validateGossipEndpoint(
+            @Nullable final List<ServiceEndpoint> endpointLst, @NonNull final NodesConfig nodesConfig) {
+        validateFalse(endpointLst == null || endpointLst.isEmpty(), INVALID_GOSSIP_ENDPOINT);
+        validateFalse(endpointLst.size() > nodesConfig.maxGossipEndpoint(), INVALID_GOSSIP_ENDPOINT);
+        // for phase 2: The first in the list is used as the Internal IP address in config.txt,
+        // the second in the list is used as the External IP address in config.txt
+        validateFalse(endpointLst.size() < 2, INVALID_GOSSIP_ENDPOINT);
+
+        for (final var endpoint : endpointLst) {
+            validateFalse(
+                    nodesConfig.gossipFqdnRestricted() && !endpoint.domainName().isEmpty(),
+                    GOSSIP_ENDPOINT_CANNOT_HAVE_FQDN);
+            validateEndpoint(endpoint, nodesConfig);
+        }
+    }
+
+    /**
+     * Validates the service endpoint.
+     *
+     * @param endpointLst The list of ServiceEndpoint to validate
+     * @param nodesConfig The nodes configuration
+     */
+    public void validateServiceEndpoint(
+            @Nullable final List<ServiceEndpoint> endpointLst, @NonNull final NodesConfig nodesConfig) {
+        validateFalse(endpointLst == null || endpointLst.isEmpty(), INVALID_SERVICE_ENDPOINT);
+        validateFalse(endpointLst.size() > nodesConfig.maxServiceEndpoint(), INVALID_SERVICE_ENDPOINT);
+        for (final var endpoint : endpointLst) {
+            validateEndpoint(endpoint, nodesConfig);
+        }
+    }
+
+    private void validateEndpoint(@NonNull final ServiceEndpoint endpoint, @NonNull final NodesConfig nodesConfig) {
+        validateFalse(endpoint.port() == 0, INVALID_ENDPOINT);
+        validateFalse(
+                endpoint.ipAddressV4().length() == 0
+                        && endpoint.domainName().trim().isEmpty(),
+                INVALID_ENDPOINT);
+        validateFalse(
+                endpoint.ipAddressV4().length() != 0
+                        && !endpoint.domainName().trim().isEmpty(),
+                IP_FQDN_CANNOT_BE_SET_FOR_SAME_ENDPOINT);
+        validateFalse(endpoint.domainName().trim().length() > nodesConfig.maxFqdnSize(), FQDN_SIZE_TOO_LARGE);
+    }
+}
